refactor(rsc): rename module export scan
