+++ packages/plugin-rsc/src/transforms/index.ts
@@ -1,5 +1,5 @@
 export * from './hoist'
-export * from './module-exports'
+export * from './module-export-scan'
 export * from './module-export-effect'
 export * from './wrap-export'
 export * from './proxy-export'
