+++ packages/plugin-rsc/src/transforms/module-export-scan.ts
